feat: 기본 db 리스트 출력 및 디테일 링크 생성 (리뷰 메인)
